requirements.txt
rake-nltk
nltk
tweepy
textblob
s3fs
pandas
nltk
tqdm
awscli
boto3
Flask
Flask-SQLAlchemy
google-cloud-secret-manager
google-cloud-datastore
google-cloud-billing
google-cloud-storage
gcsfs
bs4
gunicorn
flask-restplus
flask-cors
Werkzeug==0.16.1
plotly
marshmallow-sqlalchemy
psycopg2-binary
